requirements.txt
pyodbc
pylint
python-dotenv
requests
pandas
xlrd
xlsxwriter
openpyxl
colorama
psycopg2